wip: show blueprint 'upgrade' buttons in blueprint config page
