+++ meteor/client/ui/Settings/Migration.tsx
@@ -12,8 +12,8 @@ import * as _ from 'underscore'
 import { EditAttribute, EditAttributeBase } from '../../lib/EditAttribute'
 import { MeteorCall } from '../../../lib/api/methods'
 import { checkForOldDataAndCleanUp } from './SystemManagement'
-import { UpgradesView } from './Upgrades'
 import { stringifyError } from '@sofie-automation/shared-lib/dist/lib/stringifyError'
+import { UpgradesView } from './Upgrades/View'
 
 interface IProps {}
 interface IState {
